+++ cli/jq/cli.clj
@@ -3,9 +3,7 @@
   (:require [clojure.java.io :as io]
             [clojure.string :as str]
             [clojure.tools.cli :as cli]
-            [jq.transducers :as jq])
-  (:import (clojure.lang IReduceInit)
-           (java.io Reader BufferedReader)))
+            [jq.transducers :as jq]))
 
 (def cli-options
   [["-c" "--[no-]compact" "compact instead of pretty-printed output." :default false]
@@ -25,32 +23,17 @@
   (println "Supported options:")
   (println summary))
 
-(defn lines-from-stdin
-  "Reads lines from the stdin and pushes them further."
-  []
-  (let [^BufferedReader rdr (BufferedReader. *in*)]
-    (reify IReduceInit
-      (reduce [_ f init]
-        (try
-          (loop [state init]
-            (if (reduced? state)
-              state
-              (if-let [line (.readLine rdr)]
-                (recur (f state line))
-                state)))
-          (finally (.close rdr)))))))
-
 (defn printer
   ([_])
   ([_ item] (println item)))
 
 (defn execute [jq-expression files opts]
-  (let [xfs [(when (seq files) (map slurp))
-             (jq/parse)
+  (let [xfs [(when (seq files) (map (fn [file] (io/reader (io/file file)))))
+             (jq/rdr->json-nodes)
              (jq/execute jq-expression)
              (if (:compact opts) (jq/serialize) (jq/pretty-print))]
         xf (apply comp (remove nil? xfs))
-        values (or (seq files) (when (.ready ^Reader *in*) (lines-from-stdin)))]
+        values (or (seq files) [*in*])]
     (transduce xf printer nil values)))
 
 (defn -main [& args]